proof_of_concepts/gcp/requirements.txt
google-cloud-iam
google-cloud-compute
google-api-python-client
google-cloud-secret-manager


